refactor: restructure PTL-SI modules
