+++ .gitignore
@@ -10,3 +10,13 @@ docs/source/auto_examples/*.md5
 docs/source/auto_examples/
 docs/source/gen_modules/
 docs/_build/
+docs/source/sg_execution_times.rst
+
+# local testing
+testing/
+
+# setuptools
+build/
+dist/
+*.egg-info/
+eggs/
