--- docs/source/sg_execution_times.rst
@@ -1,40 +1,0 @@
-
-:orphan:
-
-.. _sphx_glr_sg_execution_times:
-
-
-Computation times
-=================
-**01:53.463** total execution time for 2 files **from all galleries**:
-
-.. container::
-
-  .. raw:: html
-
-    <style scoped>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
-    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
-    </style>
-    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
-    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
-    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
-    <script type="text/javascript" class="init">
-    $(document).ready( function () {
-        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
-    } );
-    </script>
-
-  .. list-table::
-   :header-rows: 1
-   :class: table table-striped sg-datatable
-
-   * - Example
-     - Time
-     - Mem (MB)
-   * - :ref:`sphx_glr_auto_examples_feature_selection_SFS_DA.py` (``..\..\examples\feature_selection\SFS_DA.py``)
-     - 01:53.244
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_feature_selection_LassoFS.py` (``..\..\examples\feature_selection\LassoFS.py``)
-     - 00:00.219
-     - 0.0
