+++ docs/source/user_guide.rst
@@ -4,4 +4,4 @@ User guide
 In the following, we provide guidance on which functions and classes to use 
 for different problems related to selective inference and its applications in 
 machine learning. Whenever possible, we point to concrete examples in the 
-documentation, which are also available as notebooks on the PSI GitHub.
+documentation, which are also available as notebooks on the PythonSI GitHub.
