+++ examples/README.txt
@@ -1,0 +1,4 @@
+Examples gallery
+================
+
+This is a gallery of all the PSI example files.
