--- examples/anomaly_detection/index.rst
@@ -1,2 +1,0 @@
-Anomaly Detection
------------------
