+++ examples/feature_selection/LassoFS.py
@@ -5,10 +5,10 @@ Selective inference for Lasso Feature Selection
 
 # Author: Tran Tuan Kiet
 
-from si import Pipeline
-from si.feature_selection import LassoFeatureSelection
-from si import Data
-from si.test_statistics import FSTestStatistic
+from pythonsi import Pipeline
+from pythonsi.feature_selection import LassoFeatureSelection
+from pythonsi import Data
+from pythonsi.test_statistics import FSTestStatistic
 import numpy as np
 import matplotlib.pyplot as plt
