--- examples/index.rst
@@ -1,28 +1,0 @@
-Examples gallery
-================
-
-This is a gallery of all the PSI example files. 
-
-Feature Selection
----------------------
-
-.. toctree::
-    :maxdepth: 2
-    :glob:
-    :hidden:
-
-    feature_selection/*
-
-.. minigallery:: ../../examples/feature_selection/*.py
-
-Anomaly Detection
----------------------
-
-.. toctree::
-    :maxdepth: 2
-    :glob:
-    :hidden:
-
-    anomaly_detection/*
-
-.. minigallery:: ../../examples/anomaly_detection/*.py
