--- examples/transfer_learning_hdr/index.rst
@@ -1,2 +1,0 @@
-Transfer Learning for High-Dimensional Regression with Selective Inference
------------------
