+++ pythonsi/__init__.py
@@ -1,0 +1,19 @@
+from pythonsi.pipeline import Pipeline
+from pythonsi import feature_selection
+from pythonsi import domain_adaptation
+from pythonsi import transfer_learning_hdr
+from pythonsi import test_statistics
+from pythonsi.node import (
+    Data,
+)
+
+__version__ = "0.0.1.post2"
+
+__all__ = [
+    "Pipeline",
+    "feature_selection",
+    "domain_adaptation",
+    "transfer_learning_hdr",
+    "test_statistics",
+    "Data",
+]
