+++ pythonsi/domain_adaptation/__init__.py
@@ -3,7 +3,7 @@ Domain adaptation methods with selective inference.
 
 """
 
-from si.domain_adaptation.optimal_transport import OptimalTransportDA
+from pythonsi.domain_adaptation.optimal_transport import OptimalTransportDA
 
 __all__ = [
     "OptimalTransportDA",
