+++ pythonsi/feature_selection/__init__.py
@@ -5,7 +5,5 @@ Feature selection methods with selective inference.
 
 from .lasso import LassoFeatureSelection
 from .seqfs import SequentialFeatureSelection
-__all__ = [
-    "LassoFeatureSelection",
-    "SequentialFeatureSelection"
-]
+
+__all__ = ["LassoFeatureSelection", "SequentialFeatureSelection"]
