+++ pythonsi/node/__init__.py
@@ -1,0 +1,5 @@
+from pythonsi.node.data import Data
+
+__all__ = [
+    "Data",
+]
