+++ pythonsi/node/data.py
