+++ pythonsi/pipeline.py
@@ -1,18 +1,19 @@
-from si.node import Data
+from pythonsi.node import Data
 import numpy as np
 import numpy.typing as npt
-from si.util import compute_p_value
+from pythonsi.util import compute_p_value
 from typing import List
 
 
-def line_search(output_node: Data, z_min: float, z_max: float, step_size: float = 1e-4):
+def line_search(output_node: Data, z_min: float, z_max: float, step_size: float = 5e-4):
     list_intervals = []
     list_outputs = []
     z = z_min
     while z < z_max:
+        # print(f"Line search at z = {z}")
         output, _, _, interval_of_z = output_node.inference(z=z)
 
-        interval_of_z = [max(interval_of_z[0], z_min), min(interval_of_z[1], z_max)]
+        interval_of_z = [max(interval_of_z[0], z_min, z), min(interval_of_z[1], z_max)]
 
         list_intervals.append(interval_of_z)
         list_outputs.append(output)
@@ -21,11 +22,6 @@ def line_search(output_node: Data, z_min: float, z_max: float, step_size: float 
         # print(f"z: {z}, interval: {interval_of_z}, output: {output}")
 
         z = interval_of_z[1] + step_size
-
-    for i in range(len(list_intervals) - 1):
-        assert list_intervals[i][1] <= list_intervals[i + 1][0] + 1e-13, (
-            f"Intervals are overlapping in line search: {list_intervals[i]} and {list_intervals[i + 1]}"
-        )
 
     return list_intervals, list_outputs
