+++ pythonsi/util/util.py
