+++ requirements.txt
@@ -1,0 +1,5 @@
+mpmath==1.3.0
+numpy==2.2.6
+POT==0.9.5
+scikit-learn==1.7.1
+scipy==1.15.3
