--- si/__init__.py
@@ -1,17 +1,0 @@
-from si.pipeline import Pipeline
-from si import feature_selection
-from si import domain_adaptation
-from si import transfer_learning_hdr
-from si import test_statistics
-from si.node import (
-    Data,
-)
-
-__all__ = [
-    "Pipeline",
-    "feature_selection",
-    "domain_adaptation",
-    "transfer_learning_hdr",
-    "test_statistics",
-    "Data",
-]
