--- si/node/__init__.py
@@ -1,5 +1,0 @@
-from si.node.data import Data
-
-__all__ = [
-    "Data",
-]
