--- si/transfer_learning_hdr/__init__.py
@@ -1,8 +1,0 @@
-"""
-Transfer learning for high-dimensional regression with selective inference.
-
-"""
-
-from .transfer_learning_hdr import PTLSITransFusion, PTLSIOracleTransLasso
-
-__all__ = ["PTLSITransFusion", "PTLSIOracleTransLasso"]
